More coding (Joseph)
I wrote the ChampionSearch class and header files.
I wrote more in the ChampDatabase class.
I commented all of my code that I wrote today.

I left off in coding the addChamp function in ChampDatabase Class.

I also fixed some issues in the ChampSearch Class.

Note: When searching for a champion, pass in a string (only the name of
the champion) and it will return the integer index of where to find that
champion in the vector of champions.