Fixed all issues (Joseph)
I fixed the issues that we were having with the different versions of
the different files.

Now, the Team and Draft classes are outdated.

How to fix:
Either Nick needs to copy paste with what he has.
Or he can revert the revert of "11/25 Team.h Team.cpp"

or I can figure out how to revert a merged commit.