UserInput (Joseph)
Deleted the Team classes from the UserInput class.  We could use them,
but it would just be inefficient to use them in most cases.

I finished the bulk of the UserInput class.
I only have a little code in the  whatDraft method.

I also updated the UML to reflect the changes I made in the class.