I fixed the compilation errors
The project compiles now. I fixed the class redefinition problems that
we were having. I also fixed the problem with defining a ChampionSearch
object.

Need to figure out why the program isn't reading in the files. It opens
"Aatrox.txt" and then enters it. However when it checks the while loop
(!(s=="\n")), it never ends.