Found it
I figured out the problem that I was having.  I thought that when I
getline("\n") that is what I would have.  I forgot that getline gets
everything in the line and ignores the \n.

I need to fix this.  I need to add a character in the files for the
addChamp function to look for and exit the while loop.