More Debugging
I fixed the checkResp() function in UserInput. Now you can use -man and
-x.

I started on the Computer Draft problem.
I narrowed it down to an infinite loop in in the compBan function.
The checkBan's true and false values are correct. Maybe it is because
there is nothing in the banned vector.

To do:
-COMMENT ALL CODE
-UPDATE UML
-fix the problem with the computer picking (counter pick)
-get <computer draft> to work.