Worked on the Computer counter pick
Found another problem. For some reason, the .txt in is reading the first
two champions in the "bad" champion file into the primary and the
secondary roles.
--FIXED--

Also found that the computer can pick a champ that you have already
picked.
--FIXED--

Score is not working correctly.
--FIXED--

Thoughts:
-Do we want to limit the computer to one of each position? If not, most
of the "winning" teams will all be of one position. And this is also
true of the user team if they want to win. Most of the counters are the
same position. And the computer will counter that position, you counter
that champion, and the computer counters that champion. Both team won't
really be realistic. Change it so it is???

To do:
-fix the above thing?
-COMMENT ALL CODE
-UPDATE THE UML