Some output fixes
I fixed some output errors. However, I found that the computer scoring
has a fault. I don't know how to fix it.

I also fixed a problem with the computer pick. How we had it coded,
there was a scenario where a champion had the position qualifications so
it would be returned to compPick. However, it was already banned or
picked. This would result in an infinite loop picking the same champion
and being told that was banned, and then picking it again. It reminds me
of the Albert Einstein quote, "Insanity is doing something over and over
expecting a different result."

To do:
-COMMENT ALL CODE
-fix computer score bug